backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.8.0
python-dotenv==1.0.0
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
tavily-python==0.3.1
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
